Update README.md
Removed the link that was causing the issue . The issue was as we are clicking on the link. it was showing the same video which is already their in that article . So I removed the link . This help to improve user experience of that article page . article page name is " Is the finance industry using open source ? Yes. Yes it is!, with Gabriele Columbro 

Signed-off-by: Dev00-1 <devmalang.2005@gmail.com>